Added all missing files and code from class.

+++ app/src/main/java/edu/miracosta/cs134/ahernandez/model/Task.java
@@ -2,11 +2,11 @@ package edu.miracosta.cs134.ahernandez.model;
 
 public class Task {
 
-    private int mId ;
+    private long mId ;
     private String mDescription ;
     private boolean mIsDone ;
 
-    public Task(int id, String description, boolean isDone) {
+    public Task(long id, String description, boolean isDone) {
         mId = id;
         mDescription = description;
         mIsDone = isDone;
@@ -22,11 +22,11 @@ public class Task {
         this(-1,description, false) ;
     }
 
-    public int getId() {
+    public long getId() {
         return mId;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         mId = id;
     }
